siglip/requirements.txt
torch
transformers
Pillow
opencv-python
matplotlib
numpy 